API-352: move AddParent service into Catalog component
