+++ src/Pim/Bundle/ConnectorBundle/Resources/config/processors.yml
@@ -7,7 +7,6 @@ parameters:
     pim_connector.processor.denormalization.job_instance.class:        Pim\Component\Connector\Processor\Denormalization\JobInstanceProcessor
     pim_connector.processor.denormalization.product_model_loader.class: Pim\Component\Connector\Processor\Denormalization\ProductModelLoaderProcessor
     pim_connector.processor.denormalization.product_to_import.class: Pim\Component\Connector\Processor\Denormalization\Product\FindProductToImport
-    pim_connector.processor.denormalization.add_parent_to_product.class: Pim\Component\Connector\Processor\Denormalization\Product\AddParent
 
     pim_connector.processor.normalization.class:                  Pim\Component\Connector\Processor\Normalization\Processor
     pim_connector.processor.normalization.product.class:          Pim\Component\Connector\Processor\Normalization\ProductProcessor
@@ -55,7 +54,7 @@ services:
         arguments:
             - '@pim_catalog.repository.product'
             - '@pim_connector.processor.denormalization.product_to_import'
-            - '@pim_connector.processor.denormalization.add_parent_to_product'
+            - '@pim_catalog.entity_with_family_variant.add_parent_to_product'
             - '@pim_catalog.updater.product'
             - '@pim_catalog.validator.product'
             - '@akeneo_storage_utils.doctrine.object_detacher'
@@ -68,13 +67,6 @@ services:
             - '@pim_catalog.repository.product'
             - '@pim_catalog.builder.product'
             - '@pim_catalog.builder.variant_product'
-
-    pim_connector.processor.denormalization.add_parent_to_product:
-        class: '%pim_connector.processor.denormalization.add_parent_to_product.class%'
-        arguments:
-            - '@pim_catalog.repository.product_model'
-            - '@pim_catalog.entity_with_family.create_variant_product_from_product'
-            - '@event_dispatcher'
 
     pim_connector.processor.denormalization.root_product_model:
         class: '%pim_connector.processor.denormalization.product_model.class%'
